Hack: Go back after clicking on external link
Facebook uses a linker script (l.php) to redirect external links, when the user clicks on the link, the WebView goes to this blank page, the default browser loads and then the WebView stays on the blank page, giving a bad user experience when the user comes back to the WebView. This hack fixes that.